requirements.txt
paho-mqtt
influxdb-client
pandas